streamlit/requirements.txt
plotly
requests
streamlit
python-dotenv
